refactor: unify create and delete FAB functionality
This commit refactors the Floating Action Button (FAB) on the all trainings screen. It now serves a dual purpose:
- When no training items are selected, the FAB triggers navigation to the create training screen.
- When one or more training items are selected, the FAB triggers the deletion of the selected items.

This change simplifies the UI by consolidating two actions into a single context-aware button. The FAB's icon also dynamically changes to reflect its current function (add or delete).

+++ feature/all-trainings/src/main/kotlin/io/github/stslex/workeeper/feature/all_trainings/ui/mvi/handler/ClickHandler.kt
@@ -23,25 +23,24 @@ internal class ClickHandler(
         when (action) {
             is Action.Click.TrainingItemClick -> processTrainingItemClick(action)
             is Action.Click.TrainingItemLongClick -> processTrainingItemLongClick(action)
-            Action.Click.CreateButton -> processCreateButtonClick()
-            Action.Click.DeleteButton -> processDeleteButtonClick()
+            Action.Click.ActionButton -> processActionButtonClick()
         }
     }
 
-    private fun processDeleteButtonClick() {
-        launch(
-            onSuccess = {
-                updateStateImmediate {
-                    it.copy(selectedItems = persistentSetOf())
+    private fun processActionButtonClick() {
+        if (state.value.selectedItems.isEmpty()) {
+            consume(Action.Navigation.CreateTraining)
+        } else {
+            launch(
+                onSuccess = {
+                    updateStateImmediate {
+                        it.copy(selectedItems = persistentSetOf())
+                    }
                 }
+            ) {
+                repository.removeTrainings(state.value.selectedItems)
             }
-        ) {
-            repository.removeTrainings(state.value.selectedItems)
         }
-    }
-
-    private fun processCreateButtonClick() {
-        consume(Action.Navigation.CreateTraining)
     }
 
     private fun processTrainingItemClick(action: Action.Click.TrainingItemClick) {
